Update coursecontent.js
Trying to fix the bug which is as follows: When you have already viewed a module before, if you come to review it, the button that used to be Complete Module becomes Next Step. When you click on Next Step, it goes to an empty 1.2

When you do quizzes, it occasionally gives you 100% even if you got less than 100% in questions.

When you click Complete Module for 1.3 and 1.4, it automatically takes you to the quiz results page showing 100% and then saying you got 2 out of 5 or 1 out of 5 respectively, even though you have never attempted the quiz before.

Code in the relevant areas is not removed but commented out